+++ py/desiutil/install.py
@@ -26,7 +26,7 @@ try:
 except ImportError:
     from io import StringIO
 if PY3:
-    from configparser import SafeConfigParser
+    from configparser import ConfigParser as SafeConfigParser
 else:
     from ConfigParser import SafeConfigParser
 from datetime import date
@@ -59,6 +59,7 @@ known_products = {
     'imaginglss': 'https://github.com/desihub/imaginglss',
     'specex': 'https://github.com/desihub/specex',
     'speclite': 'https://github.com/dkirkby/speclite',
+    'specsim': 'https://github.com/desihub/specsim',
     'specter': 'https://github.com/desihub/specter',
     'bbspecsim': 'https://desi.lbl.gov/svn/code/spectro/bbspecsim',
     'desiAdmin': 'https://desi.lbl.gov/svn/code/tools/desiAdmin',
@@ -155,6 +156,8 @@ class DesiInstall(object):
         Holds the value of :envvar:`NERSC_HOST`, or ``None`` if not defined.
     nersc_hosts : tuple
         The list of NERSC hosts names to be used for cross-installs.
+    nersc_module_dir : str
+        The directory that contains Module directories at NERSC.
     options : argparse.Namespace
         The parsed command-line options.
     product_url : str
@@ -165,6 +168,7 @@ class DesiInstall(object):
     """
     cross_install_host = 'edison'
     nersc_hosts = ('cori', 'edison', 'datatran', 'scigate')
+    nersc_module_dir = '/project/projectdirs/desi/software/modules'
 
     def __init__(self, test=False):
         """Bare-bones initialization.
@@ -675,12 +679,12 @@ class DesiInstall(object):
         initpy_found = False
         module_method = init_modules(self.options.moduleshome, method=True)
         if module_method is None:
-            message = ("Could not find the Python file in {0}/init!".format(
+            message = ("Could initialize Modules with MODULESHOME={0}!".format(
                        self.options.moduleshome))
             log.critical(message)
             raise DesiInstallException(message)
         else:
-            log.debug("Found Modules init file in {0}.".format(
+            log.debug("Initializing Modules with MODULESHOME={0}.".format(
                       self.options.moduleshome))
             self.module = MethodType(module_method, self)
         return True
@@ -758,9 +762,14 @@ class DesiInstall(object):
             if self.nersc is None:
                 self.options.moduledir = join(self.options.root, 'modulefiles')
             else:
-                # This needs to be replaced with something that isn't
-                # hard-coded.
-                nersc_module = '/project/projectdirs/desi/software/modules'
+                if self.config is not None:
+                    if self.config.has_option("Module Processing",
+                                              'nersc_module_dir'):
+                        nersc_module = self.config.get("Module Processing",
+                                                       'nersc_module_dir')
+                else:
+                    nersc_module = self.nersc_module_dir
+                log.debug("nersc_module_dir set to {0}.".format(nersc_module))
                 self.options.moduledir = join(nersc_module, self.nersc)
             if not self.options.test:
                 if not isdir(self.options.moduledir):