+++ py/desiutil/test/test_bitmask.py
@@ -2,9 +2,11 @@
 # -*- coding: utf-8 -*-
 """Test desiutil.bitmask.
 """
-
+from __future__ import (absolute_import, division,
+                        print_function, unicode_literals)
+# The line above will help with 2to3 support.
 import unittest
-from desiutil.bitmask import BitMask, _MaskBit
+from ..bitmask import BitMask, _MaskBit
 import yaml
 
 _bitdefyaml = """\